Add PathPhotoMapCreator.
